+++ src/game/BossSpider.cpp
@@ -25,7 +25,7 @@
 #include "EnemyCentipede.h"
 
 
-BossSpider::BossSpider(float x, float y)
+BossSpider::BossSpider(double x, double y)
 :Enemy(x, y, 56.0f, 41.0f)
 {
 	setImage("spider.png");
@@ -78,15 +78,15 @@ BossSpider::BossSpider(float x, float y)
 	setCurrentAnimation("walk");
 }
 
-void BossSpider::update(float dt) {
+void BossSpider::update(double dt) {
 	if(!dying) {
 		time += dt;
 		patrolTime += dt;
 
 		//setCurrentAnimation("walk");
 		const int speed_gravity = 960;
-		const float vision_range = 320;
-		const float vision_min_range = 32;
+		const double vision_range = 320;
+		const double vision_min_range = 32;
 
 		if(facing_direction == Facing::Left) {
 			if(body->v.x > -walk_speed)
@@ -123,7 +123,7 @@ void BossSpider::draw() {
 	AnimatedActor::draw();
 }
 
-bool BossSpider::doDamage(float damage) {
+bool BossSpider::doDamage(double damage) {
 	life -= damage;
 	if(life <= 0) {
 		die();
@@ -146,8 +146,8 @@ void BossSpider::die() {
 void BossSpider::onAnimationComplete(std::string anim) {
 	//cout << "BossSpider::onAnimationComplete(\"" << anim << "\")\n";
 	if(anim == "die") {
-		float pos_x = body->p.x;
-		float pos_y = body->p.y;
+		double pos_x = body->p.x;
+		double pos_y = body->p.y;
 		destroy();
 		CollectibleEnergyBall * ball = NULL;
 		ball = new CollectibleEnergyBall(pos_x-16, pos_y-10);
