+++ src/game/Sensor.cpp
@@ -1,6 +1,6 @@
 #include "Sensor.h"
 
-Sensor::Sensor(float x, float y, float w, float h)
+Sensor::Sensor(double x, double y, double w, double h)
 :Actor(x, y, w, h, true) {
   shape->sensor = true;
 	shape->collision_type = PhysicsType::Sensor;
